Update Guess.java
Might have to tinker with this a bit more because I'm not sure how it would actually function. 
My thought is right now (as I understand it) if I were to hit 'q' or 'p' or any other letter besides 'y' it would function the same way as 'n'. 
With this modification, the user would be re-prompted if they hit any key other than y or n.